+++ nas/src/com/nas/msc/pdtm/typedivers_modeltype/ctl/TypediversModeltypeCtl.java
@@ -1,5 +1,7 @@
 package com.nas.msc.pdtm.typedivers_modeltype.ctl;
 
+import java.util.HashMap;
+
 import javax.annotation.Resource;
 import javax.servlet.http.HttpSession;
 
@@ -10,7 +12,9 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.molos.tools.common.Checker;
+import com.molos.util.PropertiesUtil;
 import com.nas.beans.TypediversModeltype;
+import com.nas.globaldef.SE;
 import com.nas.msc.basemvc.controller.NASCtl;
 import com.nas.msc.pdtm.typedivers_modeltype.service.ITypediversModeltypeService;
 
@@ -49,10 +53,21 @@ public class TypediversModeltypeCtl extends NASCtl<TypediversModeltype> {
 		return 1;
 	}
 
+	@SuppressWarnings("serial")
 	@Override
 	@RequestMapping("del_td_modeltype")
 	public @ResponseBody
 	Object delById(int id) {
+		if (service.isReference(id)) {
+			return new HashMap<String, Object>() {
+				{
+					put("errorMsg", new PropertiesUtil().readValue(
+							lang(SE.LANGUAGE),
+							"pdtm.td.model.del.reference.tip"));
+				}
+			};
+		}
+
 		service.deleteByID(id);
 		return 1;
 	}
@@ -60,7 +75,7 @@ public class TypediversModeltypeCtl extends NASCtl<TypediversModeltype> {
 	@RequestMapping("is_repeat_td_modeltype_name")
 	@ResponseBody
 	public Object isRepeat(String name) {
-		return Checker.notNullList(service.queryByParaAndVal("name", name)) ? 1
+		return Checker.nullList(service.queryByParaAndVal("name", name)) ? 1
 				: 0;
 	}
